+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Created My Courses content in Academics / Browse / Favorites panels [#2169](https://github.com/rokwire/illinois-app/issues/2169).
 
 ## [4.1.10] - 2022-08-04
 ### Added